+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.xx]-2025-02-06
+
+### Updated
+
+- Send message
+  - Updated Parameters:
+  - `channel_name` (required if `is_update` is false)
+  - `recipient_id` (required)
+  - `message.body` (required)
+  - Optional: `order_id`, `event_id`, `files`, `meta`
+
 ## [1.0.0-alpha.29]-2025-02-05
 
 ### Added